B"H undo changes to net_io.c
The changes here to net_io.c are not critical, since part of the web
interface updating was to make it more robust and resilient to
missing/absent info.

However, the functionality added to net_io.c is still useful, since more
information can be deduced (for example, whether the plane is on the
ground).

It might be useful to separately commit/push the changes to net_io,
since in any case the modifications are useful for other things besides
the web ui.