B"H dump1090: use default values in macros in help text where possible
In the --help text, this commit uses default values already defined
by macros, where practical, instead of writing default values
straight into the help text.

This assist in keeping these values correct, as the help text is
automatically in sync with the default value the code is actually
using.

Only values which the preprocessor can transform to string literals
on its own have their values embedded into the help text;
therefore, values which require arithmetic, etc., to be performed
on the format stored in the macros in order to represent the
equivalent input value, are not made use of here.

Note, that some of the default values shown in the help text, may have
changed from before (for example, --ppm and --interactive-rows in my
experience), and may mean that the currently used default values are not
the ones intended. :-)